+++ src/base/b-list/b-list.ts
@@ -43,17 +43,22 @@ export const
 
 export default class bList extends iData implements iVisible, iWidth {
 	/**
+	 * Type: list of component items
+	 */
+	readonly Items!: Items;
+
+	/**
 	 * Initial list of component items
 	 */
 	@prop(Array)
-	readonly itemsProp: Items = [];
+	readonly itemsProp: this['Items'] = [];
 
 	/**
 	 * @deprecated
 	 * @see [[bList.itemsProp]]
 	 */
 	@prop({type: Array, required: false})
-	readonly valueProp?: Items;
+	readonly valueProp?: this['Items'];
 
 	/**
 	 * Initial component active value/s.
@@ -88,15 +93,15 @@ export default class bList extends iData implements iVisible, iWidth {
 	 * @see [[bList.itemsProp]]
 	 */
 	@computed({dependencies: ['value', 'itemsStore', 'deprecated']})
-	get items(): Items {
-		return <Items>this.field.get(this.deprecated ? 'value' : 'itemsStore');
+	get items(): this['Items'] {
+		return <this['Items']>this.field.get(this.deprecated ? 'value' : 'itemsStore');
 	}
 
 	/**
 	 * Sets a new list of component items
 	 * @see [[bList.itemsProp]]
 	 */
-	set items(value: Items) {
+	set items(value: this['Items']) {
 		this.field.set(this.deprecated ? 'value' : 'itemsStore', value);
 	}
 
@@ -112,7 +117,7 @@ export default class bList extends iData implements iVisible, iWidth {
 		return val != null ? o.normalizeItems(val) : val;
 	}))
 
-	value?: Items;
+	value?: this['Items'];
 
 	/**
 	 * Component active value.
@@ -153,7 +158,7 @@ export default class bList extends iData implements iVisible, iWidth {
 		return o.normalizeItems(val);
 	}))
 
-	protected itemsStore!: Items;
+	protected itemsStore!: this['Items'];
 
 	/**
 	 * True, if the component works with the deprecated API
@@ -391,7 +396,7 @@ export default class bList extends iData implements iVisible, iWidth {
 	}
 
 	/** @override */
-	protected initRemoteData(): CanUndef<CanPromise<Items | Dictionary>> {
+	protected initRemoteData(): CanUndef<CanPromise<this['Items'] | Dictionary>> {
 		if (!this.db) {
 			return;
 		}
@@ -404,7 +409,7 @@ export default class bList extends iData implements iVisible, iWidth {
 		}
 
 		if (Object.isArray(val)) {
-			this.items = this.normalizeItems(<Items>val);
+			this.items = this.normalizeItems(<this['Items']>val);
 		}
 
 		return this.items;
@@ -460,9 +465,9 @@ export default class bList extends iData implements iVisible, iWidth {
 	 * Normalizes the specified items and returns it
 	 * @param items
 	 */
-	protected normalizeItems(items: CanUndef<Items>): Items {
+	protected normalizeItems(items: CanUndef<this['Items']>): this['Items'] {
 		const
-			res = <Items>[];
+			res = <this['Items']>[];
 
 		if (!items) {
 			return res;
@@ -503,7 +508,7 @@ export default class bList extends iData implements iVisible, iWidth {
 	 * @see [[bList.normalizeItems]]
 	 */
 	@deprecated({renamedTo: 'normalizeItems'})
-	protected normalizeOptions(items: CanUndef<Items>): Items {
+	protected normalizeOptions(items: CanUndef<this['Items']>): this['Items'] {
 		return this.normalizeItems(items);
 	}
 
@@ -551,10 +556,10 @@ export default class bList extends iData implements iVisible, iWidth {
 	 *
 	 * @param value
 	 * @param oldValue
-	 * @emits `itemsChange(value: Items)`
+	 * @emits `itemsChange(value: this['Items'])`
 	 */
 	@watch(['value', 'itemsStore'])
-	protected syncItemsWatcher(value: Items, oldValue: Items): void {
+	protected syncItemsWatcher(value: this['Items'], oldValue: this['Items']): void {
 		if (!Object.fastCompare(value, oldValue)) {
 			this.initComponentValues();