+++ CHANGELOG.md
@@ -28,6 +28,7 @@ _Note: Gaps between patch versions are faulty, broken or test releases._
 #### :rocket: New Feature
 
 * Added support of method invoking `super/i-block/modules/state/set`
+* Added a new associated type: `Items` `bList`
 * Added support of interpolation of a data provider response:
   * `iInput`
   * `bList`